refactor(lexer): split `regex_div` tests into separate file (#26558)
Pure refactor. No code alterations, just moving things around.

Split `regex_div`'s tests into a separate module. This reduces the size of the main file, and is a little faster to build, as the `tests.rs` module doesn't get parsed unless compiling tests.
